pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "docchex"
description = ""
authors = [{name = "Ashkan Shokri", email = "ashkan.shokri@gmail.com"}]
license = "ISC"
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.10"
keywords = []
dynamic = ["version"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "pdfplumber>=0.11",
    "pyyaml>=6.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.urls]
Homepage = "https://ashkans.github.io/docchex"
Documentation = "https://ashkans.github.io/docchex"
Changelog = "https://ashkans.github.io/docchex/changelog"
Repository = "https://github.com/ashkans/docchex"
Issues = "https://github.com/ashkans/docchex/issues"
Discussions = "https://github.com/ashkans/docchex/discussions"
Gitter = "https://gitter.im/docchex/community"
Funding = "https://github.com/sponsors/ashkans"

[project.scripts]
docchex = "docchex.cli:main"

[tool.pdm.version]
source = "call"
getter = "scripts.get_version:get_version"

[tool.pdm.build]
# Include as much as possible in the source distribution, to help redistributors.
includes = ["src/docchex/_internal/rules/presets/*.yaml"]
excludes = ["**/.pytest_cache"]
source-includes = [
    "config",
    "docs",
    "scripts",
    "share",
    "tests",
    "duties.py",
    "zensical.toml",
    "*.md",
    "LICENSE",
]

[tool.pdm.build.wheel-data]
# Manual pages can be included in the wheel.
# Depending on the installation tool, they will be accessible to users.
# pipx supports it, uv does not yet, see https://github.com/astral-sh/uv/issues/4731.
data = [
    {path = "share/**/*", relative-to = "."},
]

[dependency-groups]
maintain = [
    "git-changelog>=2.5",
    "twine>=5.1",
    "yore>=0.3.3",
]
ci = [
    "duty>=1.6",
    "griffe>=2.0",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-randomly>=3.15",
    "pytest-xdist>=3.6",
    "ruff>=0.4",
    "ty>=0.0.14",
    "types-markdown>=3.6",
    "types-pyyaml>=6.0",
]
 docs = [
    "ghp-import>=2.1",
    "markdown-callouts>=0.4",
    "markdown-exec>=1.8",
    "mkdocstrings[python]>=0.29",
    # YORE: EOL 3.10: Remove line.
    "tomli>=2.0; python_version < '3.11'",
    "zensical>=0.0.21",
]

[tool.uv]
default-groups = ["maintain", "ci", "docs"]
